requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
aiohttp==3.9.1
gunicorn==21.2.0
